fix: add authorization header & delete: && on guard

+++ next-10/app/admin/posts/[id]/page.tsx
@@ -10,16 +10,19 @@ import {
   GetPostResponse,
   PostWithCategories,
 } from "@/app/_types/posts";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
 
 export default function Page() {
   const router = useRouter();
   const { id } = useParams();
   const [postData, setPostData] = useState<PostWithCategories | null>(null);
+  const { token } = useSupabaseSession();
 
   const onSubmitHandle = async (
     data: PostFormData,
     reset: () => void,
   ): Promise<void> => {
+    if (!token) return;
     try {
       const body: UpdatePostRequest = {
         title: data.title,
@@ -30,7 +33,7 @@ export default function Page() {
 
       await fetch(`/api/admin/posts/${id}`, {
         method: "PUT",
-        headers: { "Content-Type": "application/json" },
+        headers: { "Content-Type": "application/json", Authorization: token },
         body: JSON.stringify(body),
       });
 
@@ -43,9 +46,12 @@ export default function Page() {
   };
 
   const onSubmitDeleteHandle = async (reset: () => void): Promise<void> => {
+    if (!token) return;
+
     try {
       await fetch(`/api/admin/posts/${id}`, {
         method: "DELETE",
+        headers: { "Content-Type": "application/json", Authorization: token },
       });
 
       alert("削除しました。");
@@ -57,16 +63,18 @@ export default function Page() {
   };
 
   useEffect(() => {
-    if (!id) return;
+    if (!token) return;
 
     const fetcher = async () => {
-      const res = await fetch(`/api/admin/posts/${id}`);
+      const res = await fetch(`/api/admin/posts/${id}`, {
+        headers: { Authorization: token },
+      });
       const data: GetPostResponse = await res.json();
       setPostData(data.post);
     };
 
     fetcher();
-  }, [id]);
+  }, [id, token]);
 
   if (postData === null) return <span>Loading...</span>;
